parser-xml-valgrind: separate module for ValgrindTreeDecoder
